+++ resources/assets/js/components/enso/vuedatatable/TableCell.vue
@@ -2,6 +2,7 @@
 
     <span :class="{ 'is-clickable': column.meta.clickable }"
         @click="$emit('clicked')">
+        <slot name="hidden-controls" v-if="hiddenControls"></slot>
         <span v-if="column.meta.boolean"
             class="tag is-table-tag"
             :class="value ? 'is-success' : 'is-danger'">
@@ -14,6 +15,12 @@
         </span>
         <span v-else-if="column.meta.render">
             <slot name="custom-render"></slot>
+        </span>
+        <span v-else-if="column.meta.slot">
+            <slot :name="column.name"
+                :column="column"
+                :value="value">
+            </slot>
         </span>
         <span v-else-if="column.meta.translation">{{ i18n(value) }}</span>
         <span v-else>{{ value }}</span>
@@ -38,6 +45,10 @@ export default {
         i18n: {
             type: Function,
             required: true,
+        },
+        hiddenControls: {
+            type: Boolean,
+            default: false,
         },
     },
 };