+++ resources/assets/js/components/enso/vuedatatable/TableHeader.vue
@@ -6,26 +6,28 @@
                 v-if="template.crtNo">
                 {{ i18n(template.labels.crtNo) }}
             </th>
-            <th class="vue-table-header"
-                :class="template.align"
+            <th :class="['vue-table-header', template.align]"
                 v-for="column in template.columns"
                 :key="column.label"
-                v-if="column.meta.visible && !column.meta.hidden">
-                <span>
-                    {{ i18n(column.label) }}
-                    <span class="table-header-controls">
-                        <span class="sorter"
-                            @click="toggleSort($event, column)"
-                            v-if="column.meta.sortable">
-                            <span class="icon is-small">
-                                <fa :icon="sortIcon(column.meta.sort)" size="xs"></fa>
-                            </span>
-                        </span>
-                        <a class="delete is-small"
-                            v-if="column.meta.sort"
-                            @click="clearColumnSort(column)">
-                        </a>
+                v-if="column.meta.visible && !column.meta.hidden && !column.meta.rogue">
+                {{ i18n(column.label) }}
+                <span class="table-header-controls">
+                    <span class="icon is-small has-text-info"
+                        v-if="column.tooltip"
+                        v-tooltip="column.tooltip">
+                        <fa icon="info" size="xs"></fa>
                     </span>
+                    <span class="sorter"
+                        @click="toggleSort($event, column)"
+                        v-if="column.meta.sortable">
+                        <span class="icon is-small">
+                            <fa :icon="sortIcon(column.meta.sort)" size="xs"></fa>
+                        </span>
+                    </span>
+                    <a class="delete is-small"
+                        v-if="column.meta.sort"
+                        @click="clearColumnSort(column)">
+                    </a>
                 </span>
             </th>
             <th :class="template.align"
@@ -39,14 +41,18 @@
 
 <script>
 
+import { VTooltip } from 'v-tooltip';
 import fontawesome from '@fortawesome/fontawesome';
-import { faSort, faSortUp, faSortDown, faPlus, faFileExcel }
+
+import { faSort, faSortUp, faSortDown, faPlus, faFileExcel, faInfo }
     from '@fortawesome/fontawesome-free-solid/shakable.es';
 
-fontawesome.library.add(faSort, faSortUp, faSortDown, faPlus, faFileExcel);
+fontawesome.library.add(faSort, faSortUp, faSortDown, faPlus, faFileExcel, faInfo);
 
 export default {
     name: 'Header',
+
+    directives: { tooltip: VTooltip },
 
     props: {
         template: {