+++ resources/lang/hu/auth.php
@@ -1,0 +1,19 @@
+<?php
+
+return [
+
+    /*
+    |--------------------------------------------------------------------------
+    | Authentication Language Lines
+    |--------------------------------------------------------------------------
+    |
+    | The following language lines are used during authentication for various
+    | messages that we need to display to the user. You are free to modify
+    | these language lines according to your application's requirements.
+    |
+     */
+
+    'failed'   => 'Ezek a bejelentkezési adatok nem egyeznek a mi adatainkkal.',
+    'throttle' => 'Túl sok próbálkozás történt. Kérem próbálja újra :seconds másodperc múlva.',
+
+];