+++ resources/lang/hu/passwords.php
@@ -1,0 +1,22 @@
+<?php
+
+return [
+
+    /*
+    |--------------------------------------------------------------------------
+    | Password Reset Language Lines
+    |--------------------------------------------------------------------------
+    |
+    | The following language lines are the default lines which match reasons
+    | that are given by the password broker for a password update attempt
+    | has failed, such as for an invalid token or invalid new password.
+    |
+    */
+
+    'password' => 'A jelszónak minimum 6 karakter hosszúnak kell lennie és a megerősítéssel egyeznie kell.',
+    'reset'    => 'Jelszava alaphelyzetbe állítva!',
+    'sent'     => 'Emailben elküldtük a jelszavát alaphelyzetbe állító linket.',
+    'token'    => 'A jelszó alaphelyzetbe állító token nem érvényes.',
+    'user'     => "Ilyen email címmel rendelkező felhasználó nem található.",
+
+];